Complete day 10 part 1
Including a lot of changes to the build system so it can support multiple files (and thus multiple modules). I also removed the coverage report since it was erroring and I didn’t need it.

I added an extra level of nesting to the tests – which requires only using `>:::` with a list again – so that I could separate my tests from different modules, and have them all run in one test suite.