Improve day 10 OCaml project layout, build system
I moved the code into src/ and test/ folders. As well as allowing the code to be browsed separately from other project files, it allows me to change the Makefile to only build targets when their dependencies are out of date, instead of all the time like I was forced to do before. I removed some out-of-date arguments from the `clean` target in the Makefile, too.

I also simplified the main program to remove the commented-out alternative for day 2. I think it’s such a trivial change that both possible versions of the program don’t need to be referenced in the code.